WAbstractBackend: Update the 'volume' from percent to value

+++ src/SkGui/src/declarative/WDeclarativePlayer_p.h
@@ -90,7 +90,7 @@ public: // Variables
 
     qreal speed;
 
-    int volume;
+    qreal volume;
 
     bool shuffle;
